+++ src/main/java/com/tnki/core/memox/repository/UserDailyCheckInRecordRepository.java
@@ -5,7 +5,7 @@ import com.tnki.core.memox.model.UserDailyCheckInRecord;
 import java.util.Date;
 
 public interface UserDailyCheckInRecordRepository {
-    UserDailyCheckInRecord find(int userID, Date date);
-    void add(int userID, Date date);
+    UserDailyCheckInRecord find(long userID, Date date);
+    void add(long userID, Date date);
     void update(UserDailyCheckInRecord userDailyCheckInRecord);
 }
